refactor have minreq and minreq_test all happen in the minreq.py script and update cirrus CI config accordingly
